+++ src/main/java/insane96mcp/iguanatweaksreborn/module/movement/stamina/IStaminaModifier.java
@@ -1,6 +1,0 @@
-package insane96mcp.iguanatweaksreborn.module.movement.stamina;
-
-public interface IStaminaModifier {
-    float consumedStaminaModifier(int amplifier);
-    float regenStaminaModifier(int amplifier);
-}