+++ changelog.md
@@ -1,3 +1,5 @@
 # Changelog
 
 ## 4.0.0
+* Added config option to disable the shorter world border
+* Fixed End music being changed even if the Sound feature was disabled